+++ lib/components/private/Field/Field.tsx
@@ -7,7 +7,7 @@ import { FieldLabel } from '../../FieldLabel/FieldLabel';
 import { FieldMessage } from '../../FieldMessage/FieldMessage';
 import { FieldOverlay } from '../FieldOverlay/FieldOverlay';
 import styles from './Field.css.js';
-import { useText } from '../../../hooks/useTypography';
+import { useText } from '../../../hooks/typography';
 
 type FormElementProps = AllHTMLAttributes<HTMLFormElement>;
 export interface FieldProps {
